added c solvers (including those which does not interpolate data

+++ arttools/caldb.py
@@ -392,7 +392,7 @@ def get_ayut_inversed_psf_data_packed():
 
 @lru_cache(maxsize=1)
 def get_ayut_inverse_psf_datacube_packed():
-    ipsf = np.copy(get_ayut_inversed_psf_data_packed()[2].data)
+    ipsf = np.copy(get_ayut_inversed_psf_data_packed()[2].data).astype(float)
     #ipsf = fits.getdata(os.path.join(ARTCALDBPATH, "iPSF_hybrid.fits.gz"), 2)
     #ipsf = pickle.load(open("/srg/a1/work/srg/ARTCALDB/caldb_files/iPSF_marshall.pkl", "rb"))
     return ipsf
